removing cyclic dependency between shared_data_struct, mc_util and base_util
removing cyclic dependency shared_data_struct - base_util
