+++ agents/craftassist/tests/fake_agent.py
@@ -9,11 +9,10 @@ from typing import List, Tuple
 from droidlet.lowlevel.minecraft.mc_util import XYZ, IDM, Block
 from .utils import Look, Pos, Item, Player
 from agents.loco_mc_agent import LocoMCAgent
-from droidlet.base_util import TICKS_PER_SEC
 from droidlet.memory.craftassist.mc_memory import MCAgentMemory
 from droidlet.memory.craftassist.mc_memory_nodes import VoxelObjectNode
 from agents.craftassist.craftassist_agent import CraftAssistAgent
-from droidlet.shared_data_structs import Time, MockOpt
+from droidlet.shared_data_structs import TICKS_PER_SEC, Time, MockOpt
 from droidlet.dialog.dialogue_manager import DialogueManager
 from droidlet.dialog.map_to_dialogue_object import DialogueObjectMapper
 from droidlet.lowlevel.minecraft.shapes import SPECIAL_SHAPE_FNS
