+++ droidlet/lowlevel/minecraft/mc_util.py
@@ -28,18 +28,6 @@ class MCTime(Time):
         return self.get_world_time() / TICKS_PER_MC_DAY
 
 
-def adjacent(p):
-    """Return the positions adjacent to position p"""
-    return (
-        (p[0] + 1, p[1], p[2]),
-        (p[0] - 1, p[1], p[2]),
-        (p[0], p[1] + 1, p[2]),
-        (p[0], p[1] - 1, p[2]),
-        (p[0], p[1], p[2] + 1),
-        (p[0], p[1], p[2] - 1),
-    )
-
-
 def cluster_areas(areas):
     """Cluster a list of areas so that intersected ones are unioned
